Implemented partially writeZeroCell, issue with links

+++ src/main/java/cz/routing/filebuilder/model/TurnTableData.java
@@ -7,6 +7,6 @@ import lombok.Value;
  */
 @Value
 public class TurnTableData {
-    long turnTableId;
+    int turnTableId;
     double[][] matrix;
 }
